pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "groq>=0.34.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
]
